+++ notebooks/ggponc2tui.csv
@@ -1,0 +1,128 @@
+Group,Group Long,TUI,Type Name,Diagnosis_or_Pathology,Other_Finding,Clinical_Drug,Nutrient_or_Body_Substance,External_Substance,Therapeutic,Diagnostic
+ACTI,Activities & Behaviors,T052,Activity,x,x,,,,x,x
+ACTI,Activities & Behaviors,T053,Behavior,x,x,,,,,
+ACTI,Activities & Behaviors,T056,Daily or Recreational Activity,x,x,,,,,
+ACTI,Activities & Behaviors,T051,Event,x,x,,,,,
+ACTI,Activities & Behaviors,T064,Governmental or Regulatory Activity,x,x,,,,x,x
+ACTI,Activities & Behaviors,T055,Individual Behavior,x,x,,,,,
+ACTI,Activities & Behaviors,T066,Machine Activity,x,x,,,,x,x
+ACTI,Activities & Behaviors,T057,Occupational Activity,x,x,,,,x,x
+ACTI,Activities & Behaviors,T054,Social Behavior,x,x,,,,,
+ANAT,Anatomy,T017,Anatomical Structure,x,x,,x,,,
+ANAT,Anatomy,T029,Body Location or Region,x,x,,x,,,
+ANAT,Anatomy,T023,"Body Part, Organ, or Organ Component",x,x,,x,,,
+ANAT,Anatomy,T030,Body Space or Junction,x,x,,x,,,
+ANAT,Anatomy,T031,Body Substance,x,x,,x,,,
+ANAT,Anatomy,T022,Body System,x,x,,x,,,
+ANAT,Anatomy,T025,Cell,x,x,,x,,,
+ANAT,Anatomy,T026,Cell Component,x,x,,x,,,
+ANAT,Anatomy,T018,Embryonic Structure,x,x,,x,,,
+ANAT,Anatomy,T021,Fully Formed Anatomical Structure,x,x,,x,,,
+ANAT,Anatomy,T024,Tissue,x,x,,x,,,
+CHEM,Chemicals & Drugs,T116,"Amino Acid, Peptide, or Protein",,,x,x,x,x,x
+CHEM,Chemicals & Drugs,T195,Antibiotic,,,x,x,x,x,x
+CHEM,Chemicals & Drugs,T123,Biologically Active Substance,x,x,x,x,x,x,x
+CHEM,Chemicals & Drugs,T122,Biomedical or Dental Material,x,x,x,x,x,x,x
+CHEM,Chemicals & Drugs,T103,Chemical,x,x,x,x,x,x,x
+CHEM,Chemicals & Drugs,T120,Chemical Viewed Functionally,x,x,x,x,x,x,x
+CHEM,Chemicals & Drugs,T104,Chemical Viewed Structurally,x,x,x,x,x,x,x
+CHEM,Chemicals & Drugs,T200,Clinical Drug,x,x,x,x,x,x,x
+CHEM,Chemicals & Drugs,T196,"Element, Ion, or Isotope",x,x,x,x,x,x,x
+CHEM,Chemicals & Drugs,T126,Enzyme,x,x,x,x,x,x,x
+CHEM,Chemicals & Drugs,T131,Hazardous or Poisonous Substance,x,x,x,x,x,x,x
+CHEM,Chemicals & Drugs,T125,Hormone,x,x,x,x,x,x,x
+CHEM,Chemicals & Drugs,T129,Immunologic Factor,x,x,x,x,x,x,x
+CHEM,Chemicals & Drugs,T130,"Indicator, Reagent, or Diagnostic Aid",x,x,x,x,x,x,x
+CHEM,Chemicals & Drugs,T197,Inorganic Chemical,x,x,x,x,x,x,x
+CHEM,Chemicals & Drugs,T114,"Nucleic Acid, Nucleoside, or Nucleotide",x,x,x,x,x,x,x
+CHEM,Chemicals & Drugs,T109,Organic Chemical,x,x,x,x,x,x,x
+CHEM,Chemicals & Drugs,T121,Pharmacologic Substance,x,x,x,x,x,x,x
+CHEM,Chemicals & Drugs,T192,Receptor,x,x,x,x,x,x,x
+CHEM,Chemicals & Drugs,T127,Vitamin,x,x,x,x,x,x,x
+CONC,Concepts & Ideas,T185,Classification,,x,,,,,
+CONC,Concepts & Ideas,T077,Conceptual Entity,,x,,,,,
+CONC,Concepts & Ideas,T169,Functional Concept,,x,,,,,
+CONC,Concepts & Ideas,T102,Group Attribute,,x,,,,,
+CONC,Concepts & Ideas,T078,Idea or Concept,,x,,,,,
+CONC,Concepts & Ideas,T170,Intellectual Product,,,,,,,
+CONC,Concepts & Ideas,T171,Language,,x,,,,,
+CONC,Concepts & Ideas,T080,Qualitative Concept,,x,,,,,
+CONC,Concepts & Ideas,T081,Quantitative Concept,,x,,,,,
+CONC,Concepts & Ideas,T089,Regulation or Law,,,,,,,
+CONC,Concepts & Ideas,T082,Spatial Concept,,x,,,,,
+CONC,Concepts & Ideas,T079,Temporal Concept,,x,,,,,
+DEVI,Devices,T203,Drug Delivery Device,,,x,,,x,x
+DEVI,Devices,T074,Medical Device,,,x,,,x,x
+DEVI,Devices,T075,Research Device,,,x,,,x,x
+DISO,Disorders,T020,Acquired Abnormality,x,x,,,,,
+DISO,Disorders,T190,Anatomical Abnormality,x,x,,,,,
+DISO,Disorders,T049,Cell or Molecular Dysfunction,x,x,,,,,
+DISO,Disorders,T019,Congenital Abnormality,x,x,,,,,
+DISO,Disorders,T047,Disease or Syndrome,x,x,,,,,
+DISO,Disorders,T050,Experimental Model of Disease,x,x,,,,,
+DISO,Disorders,T033,Finding,x,x,,,,,
+DISO,Disorders,T037,Injury or Poisoning,x,x,,,x,,
+DISO,Disorders,T048,Mental or Behavioral Dysfunction,x,x,,,,,
+DISO,Disorders,T191,Neoplastic Process,x,x,,,,,
+DISO,Disorders,T046,Pathologic Function,x,x,,,,,
+DISO,Disorders,T184,Sign or Symptom,x,x,,,,,
+GENE,Genes & Molecular Sequences,T087,Amino Acid Sequence,x,x,x,x,x,,
+GENE,Genes & Molecular Sequences,T088,Carbohydrate Sequence,x,x,x,x,x,,
+GENE,Genes & Molecular Sequences,T028,Gene or Genome,x,x,x,x,x,,
+GENE,Genes & Molecular Sequences,T085,Molecular Sequence,x,x,x,x,x,,
+GENE,Genes & Molecular Sequences,T086,Nucleotide Sequence,x,x,x,x,x,,
+GEOG,Geographic Areas,T083,Geographic Area,x,x,x,x,x,,
+LIVB,Living Beings,T100,Age Group,,x,,,,,
+LIVB,Living Beings,T011,Amphibian,,x,,,,,
+LIVB,Living Beings,T008,Animal,,x,,,,,
+LIVB,Living Beings,T194,Archaeon,,x,,,,,
+LIVB,Living Beings,T007,Bacterium,x,x,,x,,,
+LIVB,Living Beings,T012,Bird,,x,,,,,
+LIVB,Living Beings,T204,Eukaryote,x,x,,x,,,
+LIVB,Living Beings,T099,Family Group,,x,,,,,
+LIVB,Living Beings,T013,Fish,,x,,,,,
+LIVB,Living Beings,T004,Fungus,x,x,,x,,,
+LIVB,Living Beings,T096,Group,,x,,,,,
+LIVB,Living Beings,T016,Human,,x,,,,,
+LIVB,Living Beings,T015,Mammal,,x,,,,,
+LIVB,Living Beings,T001,Organism,,x,,,,,
+LIVB,Living Beings,T101,Patient or Disabled Group,x,x,,,,,
+LIVB,Living Beings,T002,Plant,,x,,,,,
+LIVB,Living Beings,T098,Population Group,,x,,,,,
+LIVB,Living Beings,T097,Professional or Occupational Group,,x,,,,x,x
+LIVB,Living Beings,T014,Reptile,,x,,,,,
+LIVB,Living Beings,T010,Vertebrate,,x,,,,,
+LIVB,Living Beings,T005,Virus,x,x,,,,,
+OBJC,Objects,T071,Entity,x,x,x,x,x,x,x
+OBJC,Objects,T168,Food,x,x,x,x,x,x,x
+OBJC,Objects,T073,Manufactured Object,x,x,x,x,x,x,x
+OBJC,Objects,T072,Physical Object,x,x,x,x,x,x,x
+OBJC,Objects,T167,Substance,x,x,x,x,x,x,x
+OCCU,Occupations,T091,Biomedical Occupation or Discipline,x,x,,,,x,x
+OCCU,Occupations,T090,Occupation or Discipline,x,x,,,,x,x
+ORGA,Organizations,T093,Health Care Related Organization,,x,,,,x,x
+ORGA,Organizations,T092,Organization,,x,,,,x,x
+ORGA,Organizations,T094,Professional Society,,x,,,,x,x
+ORGA,Organizations,T095,Self-help or Relief Organization,,x,,,,x,x
+PHEN,Phenomena,T038,Biologic Function,x,x,,,,x,x
+PHEN,Phenomena,T069,Environmental Effect of Humans,x,x,,,,x,x
+PHEN,Phenomena,T068,Human-caused Phenomenon or Process,x,x,,,,x,x
+PHEN,Phenomena,T034,Laboratory or Test Result,x,x,,,,x,x
+PHEN,Phenomena,T070,Natural Phenomenon or Process,x,x,,,,x,x
+PHEN,Phenomena,T067,Phenomenon or Process,x,x,,,,x,x
+PHYS,Physiology,T043,Cell Function,x,x,,x,,,
+PHYS,Physiology,T201,Clinical Attribute,x,x,,,,,
+PHYS,Physiology,T045,Genetic Function,x,x,,x,,,
+PHYS,Physiology,T041,Mental Process,x,x,,,,,
+PHYS,Physiology,T044,Molecular Function,x,x,,x,,,
+PHYS,Physiology,T032,Organism Attribute,x,x,,,,,
+PHYS,Physiology,T040,Organism Function,x,x,,,,,
+PHYS,Physiology,T042,Organ or Tissue Function,x,x,,,,,
+PHYS,Physiology,T039,Physiologic Function,x,x,,,,,
+PROC,Procedures,T060,Diagnostic Procedure,,,,,,x,x
+PROC,Procedures,T065,Educational Activity,,,,,,x,
+PROC,Procedures,T058,Health Care Activity,,,,,,x,x
+PROC,Procedures,T059,Laboratory Procedure,,,,,,x,x
+PROC,Procedures,T063,Molecular Biology Research Technique,,,,,,x,x
+PROC,Procedures,T062,Research Activity,,,,,,x,x
+PROC,Procedures,T061,Therapeutic or Preventive Procedure,,,,,,x,x
