+++ xmen/data/__init__.py
@@ -2,3 +2,11 @@ from .indexed_dataset import IndexedDatasetDict, IndexedDataset
 from .dataloaders import *
 from .util import *
 from .integrations import *
+
+from .abbrevations import AbbreviationExpander
+from .sampling import Sampler
+from .merge_concepts import ConceptMerger
+from .retired_cuis import CUIReplacer
+from .filter import EmptyNormalizationFilter
+from .semantic_types import SemanticTypeFilter
+from .semantic_groups import SemanticGroupFilter
