+++ xmen/data/abbrevations.py
